# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Product Service API",
    "version": "1.0.0",
    "description": "Product microservice avec gestion modulaire des variants (color, size, capacity, weight) et pricing dynamique"
  },
...
  "paths": {
...
    "/api/v1/products/{id}/variants/batch": {
      "post": {
        "tags": ["Variants"],
        "summary": "Upload image and create multiple variants with modular pricing",
        "description": "Upload product image and create color/size/capacity/weight variants in batch. Supports two pricing modes: capacity-based (electronics) or modifier-based (clothing).",
        "operationId": "createBatchVariants",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image file (jpg, png, webp)"
                  },
                  "colors": {
                    "type": "string",
                    "description": "JSON array of color objects with priceImpact",
                    "example": "[{\"value\":\"#FF0000\",\"displayValue\":\"Rouge\",\"priceImpact\":50,\"type\":\"hex\"}]"
                  },
                  "sizes": {
                    "type": "string",
                    "description": "JSON array of size objects with priceImpact",
                    "example": "[{\"value\":\"XL\",\"priceImpact\":10}]"
                  },
                  "capacities": {
                    "type": "string",
                    "description": "JSON array of capacity objects with basePrice (for electronics)",
                    "example": "[{\"value\":\"256GB\",\"displayValue\":\"256 Go\",\"basePrice\":1199}]"
                  },
                  "weights": {
                    "type": "string",
                    "description": "JSON array of weight objects with priceImpact",
                    "example": "[{\"value\":0.5,\"label\":\"500g\",\"priceImpact\":5}]"
                  },
                  "basePrice": {
                    "type": "number",
                    "description": "Base price for standard products (if no capacities)",
                    "example": 29.99
                  },
                  "pricingStrategy": {
                    "type": "string",
                    "enum": ["capacity-based", "modifier-based"],
                    "default": "modifier-based",
                    "description": "capacity-based: electronics (iPhone), modifier-based: clothing"
                  },
                  "is_default": {
                    "type": "boolean",
                    "default": false,
                    "description": "Set first variant as default"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded and variants created with calculated prices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "product_id": { "type": "string", "format": "uuid" },
                        "image_url": { "type": "string" },
                        "filename": { "type": "string" },
                        "pricing_strategy": { "type": "string" },
                        "variants_created": { "type": "integer" },
                        "price_range": {
                          "type": "object",
                          "properties": {
                            "min": { "type": "number" },
                            "max": { "type": "number" }
                          }
                        },
                        "variants": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ProductVariant" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/api/v1/products/{id}/calculate-price": {
      "post": {
        "tags": ["Pricing", "Variants"],
        "summary": "Calculate price for selected variant attributes",
        "description": "Preview final price before purchase based on selected color, size, capacity with modular pricing",
        "operationId": "calculateVariantPrice",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["attributes"],
                "properties": {
                  "attributes": {
                    "type": "object",
                    "properties": {
                      "color": {
                        "type": "string",
                        "description": "Selected color value or ID",
                        "example": "#FF0000"
                      },
                      "size": {
                        "type": "string",
                        "example": "XL"
                      },
                      "capacity": {
                        "type": "string",
                        "example": "256GB",
                        "description": "For electronics"
                      },
                      "weight": {
                        "type": "number",
                        "example": 0.5
                      }
                    }
                  },
                  "quantity": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                  }
                }
              },
              "examples": {
                "clothing": {
                  "summary": "T-shirt with color and size",
                  "value": {
                    "attributes": { "color": "#000000", "size": "XL" },
                    "quantity": 2
                  }
                },
                "electronics": {
                  "summary": "iPhone with capacity and color",
                  "value": {
                    "attributes": { "capacity": "512GB", "color": "Titanium Blue" },
                    "quantity": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price calculated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "product_id": { "type": "string" },
                        "selected_attributes": {
                          "type": "object",
                          "properties": {
                            "color": { "type": "string" },
                            "size": { "type": "string" },
                            "capacity": { "type": "string" },
                            "weight": { "type": "number" }
                          }
                        },
                        "existing_variant": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "sku": { "type": "string" },
                            "stock": { "type": "integer" }
                          }
                        },
                        "price_breakdown": {
                          "type": "object",
                          "properties": {
                            "base_price": { "type": "number" },
                            "modifiers": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "type": { "type": "string", "enum": ["color", "size", "weight"] },
                                  "value": { "type": "string" },
                                  "impact": { "type": "number" }
                                }
                              }
                            },
                            "subtotal": { "type": "number" },
                            "quantity": { "type": "integer" },
                            "total": { "type": "number" }
                          }
                        },
                        "final_price": { "type": "number" },
                        "formatted_price": { "type": "string", "example": "1 299.00 €" },
                        "sku": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid attributes or combination not available" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/api/v1/products/{id}/price-matrix": {
      "get": {
        "tags": ["Pricing", "Variants"],
        "summary": "Get all price combinations matrix",
        "description": "Returns all possible variant combinations with their calculated prices. Useful for displaying a comparison grid.",
        "operationId": "getPriceMatrix",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Price matrix generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "product_id": { "type": "string" },
                        "product_name": { "type": "string" },
                        "total_combinations": { "type": "integer" },
                        "price_range": {
                          "type": "object",
                          "properties": {
                            "min": { "type": "number" },
                            "max": { "type": "number" }
                          }
                        },
                        "matrix": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "combination": {
                                "type": "object",
                                "properties": {
                                  "color": { "type": "string" },
                                  "size": { "type": "string" },
                                  "capacity": { "type": "string" }
                                }
                              },
                              "sku": { "type": "string" },
                              "price": { "type": "number" },
                              "calculated_price": { "type": "number" },
                              "override_price": { "type": "number", "nullable": true },
                              "stock": { "type": "integer" },
                              "in_stock": { "type": "boolean" },
                              "is_default": { "type": "boolean" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Product not found" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}